change the delete/insert method of table

+++ mysql.go
@@ -264,32 +264,60 @@ func (t *table) Insert(obj interface{}) (res sql.Result, err error) {
 	// obj 可以为对象
 	// 也可以为map[string]interface{}
 	if t.db != nil {
-		obv := reflect.ValueOf(obj).Elem()
-		tis := getFieldInfoFromStruct(obv)
 		insertSql := fmt.Sprintf("insert into %v(", t.name)
 		valueSql := "("
-		if len(tis) == 0 {
-			return
+
+		ro := reflect.ValueOf(obj)
+		if ro.Kind() != reflect.Ptr {
+			return nil, fmt.Errorf("only support pointer argument")
 		}
 
-		args := make([]interface{}, 0)
-		for _, v := range tis {
-			if v.fn == "_" {
-				continue
+		obv := ro.Elem()
+
+		if obv.Kind() == reflect.Struct {
+			tis := getFieldInfoFromStruct(obv)
+			if len(tis) == 0 {
+				return
+			}
+			args := make([]interface{}, 0)
+			for _, v := range tis {
+				if v.fn == "_" {
+					continue
+				}
+
+				insertSql += v.fn + ","
+				valueSql += fmt.Sprintf("?,")
+				args = append(args, v.fp.Interface())
 			}
 
-			insertSql += v.fn + ","
-			valueSql += fmt.Sprintf("?,")
-			args = append(args, v.fp.Interface())
-		}
+			if len(args) == 0 {
+				return nil, fmt.Errorf("no valid fields found in the object")
+			}
 
-		if len(args) == 0 {
-			return nil, fmt.Errorf("no valid fields found in the object")
-		}
+			sql := insertSql[0:len(insertSql)-1] + ") values" + valueSql[0:len(valueSql)-1] + ")"
+			fmt.Println("table.Insert", sql)
+			return t.db.Exec(sql, args...)
+		} else if obv.Kind() == reflect.Map {
+			keys := obv.MapKeys()
+			args := make([]interface{}, 0)
+			for _, v := range keys {
+				k := v.Interface().(string)
 
-		sql := insertSql[0:len(insertSql)-1] + ") values" + valueSql[0:len(valueSql)-1] + ")"
-		fmt.Println("table.Insert", sql)
-		return t.db.Exec(sql, args...)
+				insertSql += k + ","
+				valueSql += fmt.Sprintf("?,")
+				args = append(args, obv.MapIndex(v).Interface())
+			}
+
+			if len(args) == 0 {
+				return nil, fmt.Errorf("no valid fields found in the object")
+			}
+
+			sql := insertSql[0:len(insertSql)-1] + ") values" + valueSql[0:len(valueSql)-1] + ")"
+			fmt.Println("table.Insert", sql)
+			return t.db.Exec(sql, args...)
+		} else {
+			return nil, fmt.Errorf("non-supported input args, only supporting struct or map[string]interface{}")
+		}
 	}
 
 	return nil, fmt.Errorf("db is not opened")
@@ -299,21 +327,47 @@ func (t *table) Delete(obj interface{}) (res sql.Result, err error) {
 	// obj 可以为对象，根据pk来删除
 	// 也可以为map[string]interface{}
 	if t.db != nil {
-		obv := reflect.ValueOf(obj).Elem()
-		tis := getFieldInfoFromStruct(obv)
-		if len(tis) == 0 {
-			return
+		ro := reflect.ValueOf(obj)
+		if ro.Kind() != reflect.Ptr {
+			return nil, fmt.Errorf("only support pointer argument")
 		}
 
-		for _, v := range tis {
-			if v.pk == true {
-				//sql := fmt.Sprintf("update %v set %v=?", t.name, v.fn)
-				sql := fmt.Sprintf("delete from %v where %v=?", t.name, v.fn)
-				fmt.Println("table.Delete", sql)
-				return t.db.Exec(sql, v.fp.Interface())
+		obv := ro.Elem()
+		if obv.Kind() == reflect.Struct {
+			tis := getFieldInfoFromStruct(obv)
+			if len(tis) == 0 {
+				return
 			}
+
+			for _, v := range tis {
+				if v.pk == true {
+					sql := fmt.Sprintf("delete from %v where %v=?", t.name, v.fn)
+					fmt.Println("table.Delete", sql)
+					return t.db.Exec(sql, v.fp.Interface())
+				}
+			}
+			return nil, fmt.Errorf("no pk found")
+		} else if obv.Kind() == reflect.Map {
+			deleteSql := fmt.Sprintf("delete from %v where ", t.name)
+			keys := obv.MapKeys()
+			args := make([]interface{}, 0)
+			for _, v := range keys {
+				k := v.Interface().(string)
+
+				deleteSql += k + "=? and "
+				args = append(args, obv.MapIndex(v).Interface())
+			}
+
+			if len(args) == 0 {
+				return nil, fmt.Errorf("no valid fields found in the object")
+			}
+
+			sql := deleteSql[0 : len(deleteSql)-4]
+			fmt.Println("table.Delete", sql)
+			return t.db.Exec(sql, args...)
+		} else {
+			return nil, fmt.Errorf("non-supported input args, only supporting struct or map[string]interface{}")
 		}
-		return nil, fmt.Errorf("no pk found")
 	}
 
 	return nil, fmt.Errorf("db is not opened")
@@ -324,7 +378,12 @@ func (t *table) Update(obj interface{}) (res sql.Result, err error) {
 	// Update(key, value) key,value都是map[string]interface{}
 
 	if t.db != nil {
-		obv := reflect.ValueOf(obj).Elem()
+		ro := reflect.ValueOf(obj)
+		if ro.Kind() != reflect.Ptr {
+			return nil, fmt.Errorf("only support pointer argument")
+		}
+
+		obv := ro.Elem()
 		tis := getFieldInfoFromStruct(obv)
 		updateSql := fmt.Sprintf("update %v set ", t.name)
 		whereSql := ""
